11th june 2023

+++ src/thetestingacademy/corejava/operators_03/misc_3/Cast01.java
@@ -4,7 +4,7 @@ public class Cast01 {
     public static void main(String[] args) {
         int course = 100;
         float GST = .1845f;
-        int total = course + (int)(course*GST);
+        float total = (course + course*GST);
         System.out.println(total);
